Revert "refactoring"
This reverts commit ad41207222a83a855f7659873489e00374830d9a.

+++ DeviceManagement/DeviceManagement.Model/Device/DeviceModel.cs
@@ -27,7 +27,7 @@ namespace DeviceManagement.Model.Device
         {
             var deviceModel = new DeviceModel()
             {
-                IdDevice = (int)deviceEntity.IdDevice,
+                IdDevice = deviceEntity.IdDevice,
                 SerialNumber = deviceEntity.SerialNumber,
                 Brand = deviceEntity.Brand,
                 Model = deviceEntity.Model,
